chore: rename jax utils file
