+++ stoix/systems/ppo/ff_ppo.py
@@ -26,7 +26,7 @@ from stoix.systems.ppo.types import (
 from stoix.types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from stoix.utils import make_env as environments
 from stoix.utils.checkpointing import Checkpointer
-from stoix.utils.jax import (
+from stoix.utils.jax_utils import (
     merge_leading_dims,
     unreplicate_batch_dim,
     unreplicate_n_dims,
